requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.8.3
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.9.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.2.4
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.6.3
narwhals==2.15.0
numpy==1.26.4
packaging==25.0
pandas==2.1.4
pillow==12.1.0
propcache==0.3.2
protobuf==6.33.3
psutil==7.2.1
pyarrow==22.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.42
streamlit==1.52.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
yarl==1.20.1
